improve modulization of map_editing and coordinate_system. make cartesian coordinate_system::cartesian for extension of spherical coordinate submod
