+++ src/coordinate_system/cartesian/xzpoint.rs
@@ -1,0 +1,64 @@
+use super::xzvector::XZVector;
+use serde::Deserialize;
+use std::ops::{Add, AddAssign, Sub, SubAssign};
+
+#[derive(Debug, Deserialize, Copy, Clone)]
+pub struct XZPoint {
+    pub x: i32,
+    pub z: i32,
+}
+
+impl XZPoint {
+    #[inline]
+    pub fn new(x: i32, z: i32) -> Self {
+        Self { x, z }
+    }
+}
+
+// below are associated +- operators
+impl Add<XZVector> for XZPoint {
+    type Output = XZPoint;
+
+    fn add(self, other: XZVector) -> XZPoint {
+        XZPoint {
+            x: self.x + other.dx,
+            z: self.z + other.dz,
+        }
+    }
+}
+
+impl AddAssign<XZVector> for XZPoint {
+    fn add_assign(&mut self, other: XZVector) {
+        self.x += other.dx;
+        self.z += other.dz;
+    }
+}
+
+impl Sub for XZPoint {
+    type Output = XZVector;
+
+    fn sub(self, other: XZPoint) -> XZVector {
+        XZVector {
+            dx: self.x - other.x,
+            dz: self.z - other.z,
+        }
+    }
+}
+
+impl Sub<XZVector> for XZPoint {
+    type Output = XZPoint;
+
+    fn sub(self, other: XZVector) -> XZPoint {
+        XZPoint {
+            x: self.x - other.dx,
+            z: self.z - other.dz,
+        }
+    }
+}
+
+impl SubAssign<XZVector> for XZPoint {
+    fn sub_assign(&mut self, other: XZVector) {
+        self.x -= other.dx;
+        self.z -= other.dz;
+    }
+}
