+++ src/element_processing/landuse.rs
@@ -1,6 +1,6 @@
 use crate::args::Args;
 use crate::block_definitions::*;
-use crate::cartesian::XZPoint;
+use crate::coordinate_system::cartesian::XZPoint;
 use crate::element_processing::tree::Tree;
 use crate::floodfill::flood_fill_area;
 use crate::ground::Ground;
