+++ src/map_editing/translate/mod.rs
@@ -1,0 +1,5 @@
+mod startend_translator;
+mod translator;
+mod vector_translator;
+
+pub use translator::Translator;
