+++ src/map_editing/translate/startend_translator.rs
@@ -1,0 +1,9 @@
+use crate::coordinate_system::cartesian::XZPoint;
+use serde::Deserialize;
+
+// move the map so that start goes to end
+#[derive(Debug, Deserialize)]
+pub struct StartEndTranslator {
+    pub start: XZPoint,
+    pub end: XZPoint,
+}
